# swagger.yaml
openapi: 3.0.0
info:
  title: Nodejs Blog API
  contact: {}
  version: '1.0'
servers:
  - url: https://nodejs-blog-api-774.herokuapp.com/api/v1
    variables: {}
paths:
...
  /blog:
    post:
      tags:
        - Blog
      summary: Create Blog
      operationId: CreateBlog
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogRequest'
            example:
              title: An eye-catch blog title
              content: Blog content
              publish: true
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Blog
      summary: Get All Blogs
      operationId: GetAllBlogs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /blog/{blogId}:
    parameters:
      - in: path
        name: blogId
        schema:
          type: string
        required: true
        description: The Blog ID
    get:
      tags:
        - Blog
      summary: Get Single Blog
      operationId: GetSingleBlog
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    put:
      tags:
        - Blog
      summary: Update Blog
      operationId: UpdateBlog
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBlogRequest'
            example:
              title: An more eye-catching blog title
              content: Blog content
              publish: false

        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
        - Blog
      summary: Delete Blog
      operationId: DeleteBlog
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateBlogRequest:
      title: CreateBlogRequest
      required:
        - title
        - content
        - publish
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        publish:
          type: boolean
      example:
        title: An eye-catch blog title
        content: Blog content
        publish: true
    UpdateBlogRequest:
      title: UpdateBlogRequest
      required:
        - title
        - content
        - publish
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        publish:
          type: boolean
      example:
        title: An more eye-catching blog title
        content: Blog content
        publish: false
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...